Move entities into frontend and sort imports
