+++ src/frontend/containers/AppToolbar/FileTagEditor.tsx
@@ -9,17 +9,18 @@ import React, {
   useRef,
   useState,
 } from 'react';
-import { ClientFile } from 'src/entities/File';
-import { ClientTag } from 'src/entities/Tag';
-import { TagOption } from 'src/frontend/components/TagSelector';
-import FocusManager from 'src/frontend/FocusManager';
-import { useAction, useAutorun, useComputed } from 'src/frontend/hooks/mobx';
+
 import { debounce } from 'common/timeout';
 import { Grid, Tag } from 'widgets';
 import { Row, RowSeparator, useGridFocus } from 'widgets/Combobox/Grid';
 import { IconSet } from 'widgets/Icons';
 import { ToolbarButton } from 'widgets/menus';
+import { TagOption } from '../../components/TagSelector';
 import { useStore } from '../../contexts/StoreContext';
+import { ClientFile } from '../../entities/File';
+import { ClientTag } from '../../entities/Tag';
+import FocusManager from '../../FocusManager';
+import { useAction, useAutorun, useComputed } from '../../hooks/mobx';
 
 const POPUP_ID = 'tag-editor-popup';
