+++ src/frontend/containers/Outliner/LocationsPanel/index.tsx
@@ -3,29 +3,30 @@ import { action, runInAction } from 'mobx';
 import { observer } from 'mobx-react-lite';
 import SysPath from 'path';
 import React, { useCallback, useContext, useEffect, useMemo, useState } from 'react';
-import { ClientLocation, ClientSubLocation } from 'src/entities/Location';
-import { ClientStringSearchCriteria } from 'src/entities/SearchCriteria';
-import { LocationRemoval, SubLocationExclusion } from 'src/frontend/components/RemovalAlert';
-import { AppToaster } from 'src/frontend/components/Toaster';
-import DropContext from 'src/frontend/contexts/DropContext';
-import { useStore } from 'src/frontend/contexts/StoreContext';
-import { DnDLocationType, useLocationDnD } from 'src/frontend/contexts/TagDnDContext';
-import { useAutorun } from 'src/frontend/hooks/mobx';
-import LocationStore from 'src/frontend/stores/LocationStore';
-import { triggerContextMenuEvent } from '../utils';
-import { RendererMessenger } from 'src/ipc/renderer';
+
 import { IconSet, Tree } from 'widgets';
-import { Menu, MenuDivider, MenuItem, Toolbar, ToolbarButton, useContextMenu } from 'widgets/menus';
 import MultiSplitPane, { MultiSplitPaneProps } from 'widgets/MultiSplit/MultiSplitPane';
+import { ITreeItem, createBranchOnKeyDown } from 'widgets/Tree';
+import { Menu, MenuDivider, MenuItem, Toolbar, ToolbarButton, useContextMenu } from 'widgets/menus';
 import { Callout } from 'widgets/notifications';
-import { createBranchOnKeyDown, ITreeItem } from 'widgets/Tree';
+import { RendererMessenger } from '../../../../ipc/renderer';
+import { LocationRemoval, SubLocationExclusion } from '../../../components/RemovalAlert';
+import { AppToaster } from '../../../components/Toaster';
+import DropContext from '../../../contexts/DropContext';
+import { useStore } from '../../../contexts/StoreContext';
+import { DnDLocationType, useLocationDnD } from '../../../contexts/TagDnDContext';
+import { ClientLocation, ClientSubLocation } from '../../../entities/Location';
+import { ClientStringSearchCriteria } from '../../../entities/SearchCriteria';
+import { useAutorun } from '../../../hooks/mobx';
+import LocationStore from '../../../stores/LocationStore';
 import { IExpansionState } from '../../types';
+import { createDragReorderHelper } from '../TreeItemDnD';
 import TreeItemRevealer from '../TreeItemRevealer';
+import { triggerContextMenuEvent } from '../utils';
 import LocationCreationDialog from './LocationCreationDialog';
 import LocationRecoveryDialog from './LocationRecoveryDialog';
-import { createDragReorderHelper } from '../TreeItemDnD';
-import { useFileDropHandling } from './useFileDnD';
 import { onDragOver as onDragOverFileDnD } from './dnd';
+import { useFileDropHandling } from './useFileDnD';
 
 export class LocationTreeItemRevealer extends TreeItemRevealer {
   private locationStore?: LocationStore;
