+++ src/frontend/containers/Outliner/TagsPanel/TagMerge.tsx
@@ -1,10 +1,11 @@
 import { action } from 'mobx';
 import { observer } from 'mobx-react-lite';
 import React, { useState } from 'react';
-import { ClientTag } from 'src/entities/Tag';
-import { useStore } from 'src/frontend/contexts/StoreContext';
+
 import { Button, GridCombobox, GridOption, GridOptionCell, IconSet, Tag } from 'widgets';
 import { Dialog } from 'widgets/popovers';
+import { useStore } from '../../../contexts/StoreContext';
+import { ClientTag } from '../../../entities/Tag';
 
 interface TagMergeProps {
   tag: ClientTag;
@@ -98,10 +99,7 @@ const renderTagOption = action((tag: ClientTag, index: number, selection: boolea
   return (
     <GridOption key={id} rowIndex={index} selected={selection || undefined} tooltip={path}>
       <GridOptionCell id={id} colIndex={1}>
-        <span
-          className="combobox-popup-option-icon"
-          style={{ color: tag.viewColor }}
-        >
+        <span className="combobox-popup-option-icon" style={{ color: tag.viewColor }}>
           {IconSet.TAG}
         </span>
         {tag.name}
