+++ src/frontend/entities/SearchCriteria.ts
@@ -1,6 +1,6 @@
 import { action, Lambda, makeObservable, observable } from 'mobx';
 
-import { camelCaseToSpaced } from '../../common/fmt';
+import { camelCaseToSpaced } from '../../../common/fmt';
 import {
   ArrayConditionDTO,
   ConditionDTO,
@@ -9,9 +9,9 @@ import {
   NumberOperatorType,
   StringConditionDTO,
   StringOperatorType,
-} from '../api/data-storage-search';
-import { FileDTO } from '../api/file';
-import { ID } from '../api/id';
+} from '../../api/data-storage-search';
+import { FileDTO } from '../../api/file';
+import { ID } from '../../api/id';
 import {
   IBaseSearchCriteria,
   IDateSearchCriteria,
@@ -21,8 +21,8 @@ import {
   OperatorType,
   SearchCriteria,
   TagOperatorType,
-} from '../api/search-criteria';
-import RootStore from '../frontend/stores/RootStore';
+} from '../../api/search-criteria';
+import RootStore from '../stores/RootStore';
 
 // A dictionary of labels for (some of) the keys of the type we search for
 export type SearchKeyDict = Partial<Record<keyof FileDTO, string>>;
