+++ src/frontend/entities/Tag.ts
@@ -1,9 +1,9 @@
 import { IReactionDisposer, action, computed, makeObservable, observable, reaction } from 'mobx';
 
-import { MAX_TAG_DEPTH } from '../../common/config';
-import { ID } from '../api/id';
-import { ROOT_TAG_ID, TagDTO } from '../api/tag';
-import TagStore from '../frontend/stores/TagStore';
+import { MAX_TAG_DEPTH } from '../../../common/config';
+import { ID } from '../../api/id';
+import { ROOT_TAG_ID, TagDTO } from '../../api/tag';
+import TagStore from '../stores/TagStore';
 
 /**
  * A Tag as it is stored in the Client.
